+++ app/interfaces/Api.types.ts
@@ -1,6 +1,9 @@
 export interface ApiTypes {
+  _id: String;
   guild_id: String;
   role: String;
   api_name: String;
   api_adress: String;
+  last_ping: Boolean;
+  staty_score: Number;
 }