+++ app/routes/api.ts
@@ -50,12 +50,36 @@ route.post("/api", staty, async (req, res) => {
       role: role,
       api_name: name,
       api_adress: adress,
+      last_ping: true,
     });
     await newApi.save();
 
     res.status(200).json({ message: "New api added", data: newApi });
   } catch (error: any) {
     logs("error", "api:register_new", error, guild);
+    res.status(500).json({ message: "An error occured (catch)" });
+  }
+});
+
+route.put("/api/ping/:id", staty, async (req, res) => {
+  const { id } = req.params;
+  const { score, state } = req.body;
+
+  try {
+    const updatedApi = await Api.findByIdAndUpdate(
+      { _id: id },
+      {
+        staty_score: score,
+        last_ping: state,
+      },
+      {
+        new: true,
+      }
+    );
+
+    res.status(200).json({ message: "Api updated", data: updatedApi });
+  } catch (error: any) {
+    logs("error", "api:update:api", error);
     res.status(500).json({ message: "An error occured (catch)" });
   }
 });